Phase2 without camera controller + texture, renderState, material

+++ source/core/game_states/state_manager/state_manager.cpp
@@ -1,7 +1,6 @@
 #include "state_manager.hpp"
-#include "../i_game_state.hpp"
 
-void StateManager::runState(float deltaTime) {
+void StateManager::runState(double deltaTime) {
     if(!_isExiting){
         if(nextState!= nullptr){
             if(currentState!= nullptr){
